+++ main.py
@@ -85,7 +85,7 @@ from locales.LocaleManager import LocaleManager
 from src.backend.MediaManager import MediaManager
 from src.backend.AssetManagerBackend import AssetManagerBackend
 from src.backend.PageManagement.PageManagerBackend import PageManagerBackend
-from src.backend.SettingsManager import AppSettings, SettingsManager
+from src.backend.SettingsManager import SettingsManager
 from src.backend.PluginManager.PluginManager import PluginManager
 from src.backend.IconPackManagement.IconPackManager import IconPackManager
 from src.backend.WallpaperPackManagement.WallpaperPackManager import WallpaperPackManager
@@ -232,8 +232,7 @@ def create_global_objects():
 
 @log.catch
 def update_assets():
-    settings = gl.settings_manager.load_settings_from_file(os.path.join(gl.DATA_PATH, "settings", "settings.json"))
-    auto_update = AppSettings(settings).auto_update
+    auto_update = gl.settings_manager.app().auto_update
 
     if gl.argparser.parse_args().devel:
         auto_update = False